Keep related methods in same file
